% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_zero_tally}
\alias{check_zero_tally}
\title{Check which rows contain zero}
\usage{
check_zero_tally(x, column)
}
\arguments{
\item{x}{a dataframe}

\item{column}{the column name with tally counts (tidyeval)}
}
\value{

}
\description{
Appends a dataframe with a new column to identify rows that contain zeros.
This is then represented as a factor with the levels: "no" and "yes" so that
both levels are observed for plotting purposes
}
