% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gtable_filter_remove}
\alias{gtable_filter_remove}
\title{Remove elements from ggplot at Glob level}
\usage{
gtable_filter_remove(x, name, trim = TRUE)
}
\arguments{
\item{x}{the ggplot after being parsed as a gtable}

\item{name}{the glob element name}

\item{trim}{logical flag}
}
\value{

}
\description{
Remove elements from ggplot at Glob level
}
