% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_report.R
\name{launch_shiny_report}
\alias{launch_shiny_report}
\title{Launch the shiny version of the report}
\usage{
launch_shiny_report(connection = NULL, schema = "decovid_omop")
}
\arguments{
\item{parms_list}{list version of \code{parms} from
\code{\link{render_report}}.}
}
\value{

}
\description{
Launch the shiny version of the report
}
