% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{match_concepts}
\alias{match_concepts}
\title{Match and replace Athena codes with proper names}
\usage{
match_concepts(column, lookup)
}
\arguments{
\item{column}{an integer vector of Athena codes}

\item{lookup}{an extract of the \code{concept} table with 2 columns:
\enumerate{
\item concept_name (containing the full name replacement)
\item concept_id (containing the Athena code)
}}
}
\value{

}
\description{
Match and replace Athena codes with proper names
}
