% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mini_dict}
\alias{mini_dict}
\title{Create a mini concept dictionary}
\usage{
mini_dict(ctn, schema, concept_ids)
}
\arguments{
\item{ctn}{a database connection object}

\item{schema}{the database schema}

\item{concept_ids}{an integer vector of athena concept ids}
}
\value{

}
\description{
It is convenient to work with a small dictionary of concepts. This function
pulls them out and creates a standardised dictionary of concept ids and names
for use in other functions.
}
