% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_tm}
\alias{plot_tm}
\title{Plot Transition Matrix}
\usage{
plot_tm(tm, possible_states)
}
\arguments{
\item{tm}{a transition matrix from \code{\link{transition_matrix}}}

\item{possible_states}{a character vector of the possible states to be used
in the markov chain fitting process}
}
\value{

}
\description{
Plot Transition Matrix
}
