% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_tables.R
\name{prepare_cohort_boundaries}
\alias{prepare_cohort_boundaries}
\title{Prepare Cohort Boundaries}
\usage{
prepare_cohort_boundaries(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Prepare Cohort Boundaries
}
