% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_tables.R
\name{prepare_null_counts}
\alias{prepare_null_counts}
\title{Prepare NULL count information}
\usage{
prepare_null_counts(st)
}
\arguments{
\item{st}{a prepped table list from \code{\link{prepare_tables}}}
}
\value{

}
\description{
Prepare NULL count information
}
