% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_tables.R
\name{prepare_overview}
\alias{prepare_overview}
\title{Prepare Overview Table}
\usage{
prepare_overview(x)
}
\arguments{
\item{x}{the output from \code{\link{prepare_tables}}}
}
\value{

}
\description{
Prepare Overview Table
}
