% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarise_missing}
\alias{summarise_missing}
\title{Summarise Missing Values within a Dataframe}
\usage{
summarise_missing(x, tbl_name)
}
\arguments{
\item{x}{a dataframe}

\item{tbl_name}{the OMOP table name of \code{x}}
}
\value{

}
\description{
Summarise Missing Values within a Dataframe
}
