% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tally_concept}
\alias{tally_concept}
\title{Tally OMOP Concept ID Column}
\usage{
tally_concept(ctn, schema, table, concept_col)
}
\arguments{
\item{ctn}{a database connection object}

\item{schema}{the database schema}

\item{table}{the OMOP target table}

\item{concept_col}{the concept_id target column (but be \code{..._concept_id})}
}
\value{

}
\description{
Tally OMOP Concept ID Column
}
