% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_tables.R
\name{prepare_tally}
\alias{prepare_tally}
\title{Prepare Tally of Clinical Tables}
\usage{
prepare_tally(ctn, schema)
}
\arguments{
\item{ctn}{database connection}

\item{schema}{character vector of target schema}
}
\value{

}
\description{
Prepare Tally of Clinical Tables
}
